+++ IO/Decoding/Osu/Standard/HitObjectParser.cs
@@ -16,7 +16,7 @@ namespace PBGame.IO.Decoding.Osu.Standard
 		private int extraComboOffset;
 
 
-		public HitObjectParser(double offset, int formatVersion) : base(offset, formatVersion) {}
+		public HitObjectParser(float offset, int formatVersion) : base(offset, formatVersion) {}
 
 		protected override HitObject CreateCircle (Vector2 pos, bool isNewCombo, int comboOffset)
 		{
@@ -34,7 +34,7 @@ namespace PBGame.IO.Decoding.Osu.Standard
 		}
 
 		protected override HitObject CreateSlider (Vector2 pos, bool isNewCombo, int comboOffset, Vector2[] controlPoints,
-            double length, PathTypes pathType, int repeatCount, List<List<SoundInfo>> nodeSamples)
+            float length, PathTypes pathType, int repeatCount, List<List<SoundInfo>> nodeSamples)
 		{
 			isNewCombo |= forceNewCombo;
 			comboOffset += extraComboOffset;
@@ -52,7 +52,7 @@ namespace PBGame.IO.Decoding.Osu.Standard
 			};
 		}
 
-		protected override HitObject CreateSpinner (Vector2 pos, bool isNewCombo, int comboOffset, double endTime)
+		protected override HitObject CreateSpinner (Vector2 pos, bool isNewCombo, int comboOffset, float endTime)
 		{
 			forceNewCombo |= (formatVersion <= 8 || isNewCombo);
 			extraComboOffset += comboOffset;
@@ -63,7 +63,7 @@ namespace PBGame.IO.Decoding.Osu.Standard
 			};
 		}
 
-		protected override HitObject CreateHold (Vector2 pos, bool isNewCombo, int comboOffset, double endTime)
+		protected override HitObject CreateHold (Vector2 pos, bool isNewCombo, int comboOffset, float endTime)
 		{
 			return null;
 		}
